Refactored the following:
Renamed Layer to Face as we're only using faces when capturing (Layer may be used later)
Moved Face capturing related stuff to FaceCapturer
Commented out the ImageView that was shown when a picture is taken (we know setPolyToPoly works)
Removed some comments and unused code
